requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
apify==2.4.0
apify_client==1.9.2
apify_fingerprint_datapoints==0.0.2
apify_shared==1.3.2
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
Automat==24.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
binaryornot==0.4.4
Brotli==1.1.0
browserforge==1.2.3
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
crawlee==0.6.5
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
Django==5.1.7
django-browser-reload==1.18.0
django-tailwind==3.7.0
django-unfold==0.55.3
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.18.0
googlesearch-python==1.3.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.1
lazy-object-proxy==1.10.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
multidict==6.2.0
mysqlclient==2.2.7
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
parsel==1.10.0
playwright==1.51.0
prompt_toolkit==3.0.50
propcache==0.3.0
Protego==0.4.0
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.6.1
pydantic_core==2.27.2
PyDispatcher==2.0.7
pyee==12.1.1
Pygments==2.19.1
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
Scrapy==2.12.0
scrapy-playwright==0.0.43
selenium==4.31.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
text-unidecode==1.3
tldextract==5.1.3
tomli==2.2.1
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
w3lib==2.3.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
whitenoise==6.9.0
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2
zstandard==0.23.0
aiosmtplib==4.0.1
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.14
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0